requirements.txt
# GraphQL Framework
strawberry-graphql[debug-server]==0.278.1

# Web Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Database
sqlalchemy==2.0.43
psycopg[binary, pool]==3.2.9

# Configuration
python-dotenv==1.1.1
structlog==25.4.0
dependency-injector==4.48.1
pyjwt==2.10.1
bcrypt==4.3.0

# Development
pytest==8.4.1
